Add files via upload
The color filters have been implemented in the Single Rotation Cellular Automaton simulator. First version of the Shannon entropy numerical calculation was implemented in both simulators, using a system that allows for 8 colors to be distinguished within live cells. This allows for 9 possibles states for each cell (designed in such a way that  each one of them has a different "energy-level"). The behaviour of the entropy for different automatons has been observed and an hypothesis has been reached. "The Shannon entropy of reversible cellular automaton will always reach a steady value and remain constant after that, while it will either explode or decrease to 0 for any non reversible cellular automaton."